+++ Sofee.Core.Api.Tests.Unit/Services/Foundations/Languages/LanguageServiceTests.Validation.Add.cs
@@ -10,7 +10,7 @@ using Sofee.Core.Api.Models.Languages;
 using Sofee.Core.Api.Models.Languages.Exceptions;
 using Xunit;
 
-namespace Sofee.Core.Api.Tests.Unit.Services.Foundations
+namespace Sofee.Core.Api.Tests.Unit.Services.Foundations.Languages
 {
     public partial class LanguageServiceTests
     {
